packages/serverless/lib/plugins/aws/bedrock-agentcore/examples/strands-s3-code/requirements.txt
strands-agents>=0.1.0
bedrock-agentcore>=0.1.7
